# BFF/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Ogree BFF:",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/domains/bulk": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "An array of domains should be provided in the body.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Create multiple domains in a single request.",
        "operationId": "CreateBulkDomain",
        "parameters": [
          {
            "default": [
              {}
            ],
            "name": "body",
            "in": "body",
            "required": true
          }
        ]
      }
    },
...
    "/hierarchy/domains": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Return is arranged by relationship (father:[children]),\nstarting with \"Root\":[root domains].",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Returns domain complete hierarchy.",
        "operationId": "GetCompleteDomainHierarchy"
      }
    },
...
    "/users": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Get a list of users that the caller is allowed to see.",
        "operationId": "GetAllAccounts"
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Create an account with email and password credentials, it returns\na JWT key to use with the API.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Create a new user user.",
        "operationId": "CreateAccount",
        "parameters": [
          {
            "format": "object",
            "example": "{\"name\": \"John Doe\", \"roles\": {\"*\": \"manager\"}, \"email\": \"user@test.com\", \"password\": \"secret123\"}",
            "description": "Mandatory: email, password and roles. Optional: name. Roles is an object with domains as keys and roles as values. Possible roles: manager, user and viewer",
            "name": "body",
            "in": "body",
            "required": true
          }
        ]
      }
    },
    "/users/bulk": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Create multiples users with one request.",
        "operationId": "CreateBulk",
        "parameters": [
          {
            "format": "object",
            "example": "[{\"name\": \"John Doe\", \"roles\": {\"*\": \"manager\"}, \"email\": \"user@test.com\"}]",
            "description": "An array of users. Same mandatory and optional parameters as user apply, except for password. If not provided, one will be automatically created by the API.",
            "name": "body",
            "in": "body",
            "required": true
          }
        ]
      }
    },
...
    "/users/{userid}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Remove the specified user account.",
        "operationId": "RemoveAccount",
        "parameters": [
          {
            "type": "string",
            "example": "someUserId",
            "description": "The ID of the user to delete",
            "name": "userid",
            "in": "path",
            "required": true
          }
        ]
      },
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Organization"
        ],
        "summary": "Modify user permissions: domain and role.",
        "operationId": "ModifyUserRoles",
        "parameters": [
          {
            "type": "string",
            "example": "someUserId",
            "description": "The ID of the user to modify roles",
            "name": "userid",
            "in": "path",
            "required": true
          },
          {
            "type": "json",
            "example": "{\"roles\": {\"*\": \"manager\"}}",
            "description": "An object with domains as keys and roles as values",
            "name": "roles",
            "in": "body",
            "required": true
          }
        ]
      }
    },
...
  },
...
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}